+++ src/app/[userNickname]/[listId]/_components/ListDetailOuter/Comments.css.ts
@@ -70,49 +70,6 @@ export const totalCount = style({
   fontWeight: 600,
 });
 
-export const commentOuterWrapper = style({
-  marginBottom: '10px',
-
-  position: 'relative',
-
-  display: 'flex',
-  justifyContent: 'space-between',
-});
-
-export const commentWrapper = style({
-  display: 'flex',
-  gap: '8px',
-  alignItems: 'center',
-});
-
-export const commentContainer = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '5px',
-});
-
-export const commentInformationWrapper = style({
-  display: 'flex',
-  alignItems: 'baseline',
-  gap: '8px',
-});
-
-export const commentWriter = style({
-  fontSize: '1.2rem',
-  fontWeight: 600,
-});
-
-export const commentCreatedTime = style({
-  fontSize: '1rem',
-  fontWeight: 500,
-  color: '#494949',
-});
-
-export const commentContent = style({
-  fontSize: '1.2rem',
-  fontWeight: 500,
-});
-
 export const profileImage = style({
   width: '30px',
   minWidth: '30px',
@@ -121,48 +78,6 @@ export const profileImage = style({
 
   borderRadius: '16px',
   backgroundColor: '#909090',
-});
-
-export const repliesOuterWrapper = style({
-  display: 'flex',
-  justifyContent: 'space-between',
-  marginBottom: '20px',
-});
-
-export const repliesWrapper = style({});
-
-export const replyWrapper = style({
-  marginLeft: '30px',
-
-  display: 'flex',
-  gap: '8px',
-});
-
-export const showMoreRepliesWrapper = style({
-  marginLeft: '30px',
-  marginBottom: '25px',
-
-  display: 'flex',
-  alignItems: 'center',
-  gap: '3px',
-});
-
-export const showMoreReplies = style({
-  fontSize: '1rem',
-});
-
-export const createReplyButton = style({
-  padding: '0 0 0 36px',
-  marginBottom: '16px',
-
-  background: 'none',
-  fontSize: '1rem',
-  color: '#494949',
-  fontWeight: 500,
-});
-
-export const deleteButton = style({
-  cursor: 'pointer',
 });
 
 export const formNicknameWrapper = style({
